+++ packages/connect-wallet-modal/src/connectButtons/connectTrust.tsx
@@ -56,11 +56,7 @@ const ConnectTrust: FC<ConnectWalletProps> = (props) => {
   }, [connect, onBeforeConnect, onClickTrust, setRequirements]);
 
   return (
-    <ConnectButton
-      {...rest}
-      iconSrcOrReactElement={<WalletIcon />}
-      onClick={handleConnect}
-    >
+    <ConnectButton {...rest} icon={WalletIcon} onClick={handleConnect}>
       Trust Wallet
     </ConnectButton>
   );