+++ packages/connect-wallet-modal/src/connectButtons/connectZenGo.tsx
@@ -85,11 +85,7 @@ const ConnectZenGo: FC<ConnectWalletProps> = (props) => {
   ]);
 
   return (
-    <ConnectButton
-      {...rest}
-      iconSrcOrReactElement={<WalletIcon />}
-      onClick={handleConnect}
-    >
+    <ConnectButton {...rest} icon={WalletIcon} onClick={handleConnect}>
       ZenGo
     </ConnectButton>
   );