+++ packages/connect-wallet-modal/src/connectButtons/types.ts
@@ -1,10 +1,10 @@
-import { ReactElement } from 'react';
 import { ButtonProps } from '@reef-knot/ui-react';
 import { WalletAdapterData } from '@reef-knot/types';
 import { ButtonsCommonProps } from '../components';
 
 export type ConnectButtonProps = {
-  iconSrcOrReactElement: string | ReactElement;
+  icon: WalletAdapterData['icon'];
+  shouldInvertWalletIcon?: boolean;
 } & ButtonProps;
 
 export type ConnectWalletProps = ButtonsCommonProps & ButtonProps;