+++ packages/wallets/okx/src/custom.d.ts
@@ -1,0 +1,10 @@
+declare module '*.svg' {
+  const content: React.FunctionComponent<React.SVGAttributes<SVGElement>>;
+  export default content;
+}
+
+declare interface Window {
+  okxwallet: {
+    isOkxWallet: boolean;
+  };
+}