+++ packages/web3-react/test/hooks/useConnectors.test.tsx
@@ -10,7 +10,7 @@ import { Web3Provider } from '@ethersproject/providers';
 import { renderHook as renderHookOnServer } from '@testing-library/react-hooks/server';
 import { SafeAppConnector } from '@gnosis.pm/safe-apps-web3-react';
 import { CHAINS } from '@lido-sdk/constants';
-import { useAccount, useConnect } from 'wagmi';
+import { useAccount, useConnect, useNetwork } from 'wagmi';
 import { useAutoConnect, useConnectors, useWeb3 } from '../../src';
 import { ProviderWeb3 } from '../../src/context';
 
@@ -25,11 +25,13 @@ const mockUseAutoConnect = useAutoConnect as jest.MockedFunction<
 >;
 const mockUseConnect = useConnect as jest.MockedFunction<typeof useConnect>;
 const mockUseAccount = useAccount as jest.MockedFunction<typeof useAccount>;
+const mockUseNetwork = useNetwork as jest.MockedFunction<typeof useNetwork>;
 
 beforeEach(() => {
   mockUseAutoConnect.mockImplementation(() => true);
   mockUseConnect.mockReturnValue({ error: null } as any);
   mockUseAccount.mockReturnValue({ connector: {}, isConnected: false } as any);
+  mockUseNetwork.mockReturnValue({ chain: undefined, chains: [] } as any);
 });
 
 afterEach(() => {
@@ -38,6 +40,7 @@ afterEach(() => {
   mockUseAccount.mockReset();
   mockSafeAppConnector.mockReset();
   mockWeb3Provider.mockReset();
+  mockUseNetwork.mockReset();
 });
 
 const rpc = {