+++ packages/connect-wallet-modal/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @reef-knot/connect-wallet-modal
 
+## 1.0.8
+
+### Patch Changes
+
+- Tally => Taho in conflict checks
+- Reset the Terms acceptance for all users
+- Updated dependencies
+  - @reef-knot/web3-react@1.0.6
+
+## 1.0.7
+
+### Patch Changes
+
+- Update Terms text
+- Updated dependencies
+  - @reef-knot/ui-react@1.0.3
+
+## 1.0.6
+
+### Patch Changes
+
+- Replace Tally legacy connector with Taho wallet adapter
+- Fix: duplicated "Injected" react element key for wallet connection buttons
+- Updated dependencies
+  - @reef-knot/web3-react@1.0.5
+
 ## 1.0.5
 
 ### Patch Changes